+++ utils/scm/gitmanager.go
@@ -147,83 +147,3 @@ func getRemoteUrl(remote *goGit.Remote) (remoteUrl string, err error) {
 	}
 	return remote.Config().URLs[0], nil
 }
-
-// Normalize the URL by removing protocol prefix and any trailing ".git"
-func normalizeGitUrl(url string) string {
-	// jfrog-ignore - false positive, not used for communication
-	url = strings.TrimPrefix(url, "http://")
-	url = strings.TrimPrefix(url, "https://")
-	url = strings.TrimPrefix(url, "ssh://")
-	return strings.TrimSuffix(url, ".git")
-}
-
-func getGitRepoName(url string) string {
-	urlParts := strings.Split(normalizeGitUrl(url), "/")
-	return urlParts[len(urlParts)-1]
-}
-
-func getGitProject(url string) string {
-	// First part after base Url is the owner or the organization name.
-	urlParts := strings.Split(normalizeGitUrl(url), "/")
-	if len(urlParts) < 2 {
-		log.Debug(fmt.Sprintf("Failed to get project name from URL: %s", url))
-		return ""
-	}
-	projectPathComponents := []string{}
-	// Loop from the second element to the second last element of the URL. (first part is the base URL, last part is the repo name)
-	for i := 1; i < len(urlParts)-1; i++ {
-		if i == 1 && urlParts[i] == "scm" {
-			// In BB ssh clone url looks like this: https://git.id.info/scm/repo-name/repo-name.git --> ['git.id.info', 'scm', 'repo-name', 'repo-name']
-			continue
-		}
-		// Aws code commit clone url looks like this: https://git-codecommit.{region}.amazonaws.com/v1/repos/{repository_name} --> ['git-codecommit.{region}.amazonaws.com', 'v1', 'repos', '{repository_name}']
-		if len(urlParts) > 3 && ((i == 1 && urlParts[i] == "v1") || (i == 2 && urlParts[i] == "repos")) {
-			continue
-		}
-		projectPathComponents = append(projectPathComponents, urlParts[i])
-	}
-	if len(projectPathComponents) == 0 {
-		// In Gerrit clone URL looks like this: https://gerrit.googlesource.com/git-repo --> ['gerrit.googlesource.com', 'git-repo']
-		// add repo name (last part of the URL) as project name
-		projectPathComponents = append(projectPathComponents, urlParts[len(urlParts)-1])
-	}
-
-	return strings.Join(projectPathComponents, "/")
-}
-
-func getGitProvider(url string) ScProvider {
-	if strings.Contains(url, Github.String()) {
-		return Github
-	}
-	if strings.Contains(url, Gitlab.String()) {
-		return Gitlab
-	}
-	if isBitbucketProvider(url) {
-		return Bitbucket
-	}
-	if strings.Contains(url, Azure.String()) {
-		return Azure
-	}
-	if strings.Contains(url, Gerrit.String()) {
-		return Gerrit
-	}
-	if strings.Contains(url, Gitea.String()) {
-		return Gitea
-	}
-	if strings.Contains(url, AWSCodeCommit.String()) {
-		return AWSCodeCommit
-	}
-	// Unknown for self-hosted git providers
-	log.Debug(fmt.Sprintf("Unknown git provider for URL: %s", url))
-	return Unknown
-}
-
-func isBitbucketProvider(url string) bool {
-	if urlParts := strings.Split(normalizeGitUrl(url), "/"); len(urlParts) > 2 && urlParts[1] == "scm" {
-		return true
-	}
-	if projectName := getGitProject(url); strings.Contains(projectName, "~") {
-		return true
-	}
-	return strings.Contains(url, Bitbucket.String())
-}
